requirements.txt
numpy
scipy
matplotlib

opencv-python
open3d

torch
torchvision

pyyaml

tqdm